Add: BindedCommand
This command accepts a callback that's parameter is a <Binder>. This binder should be a class with OptionAttributes (found in CommandLine). When this command is called, the arguments will be parsed into a <Binder> and passed to your custom callback. This allows commands to have very detailed arguments & an automatic '--help'